sso_login/requirements.txt
Flask==2.1.2
Flask-OAuthlib==0.9.6
requests==2.28.1
werkzeug==2.0.3
mysql-connector-python==8.3.0
flask-cors